A user can now register into an event and his name will be displayed in the event page globally.
To test, one has to login as a user and agree to participate in event (by accesing the event), he will see the result in the same page
One also has to either manually input an event into the database or run the fixture. Registering is also necessary.

To load the fixtures: `php bin/console doctrine:fixtures:load`